mongodb-search-core/src/main/kotlin/io/github/yearnlune/search/core/operator/EqualOperator.kt
package io.github.yearnlune.search.core.operator

import org.springframework.data.mongodb.core.aggregation.AggregationExpression
import org.springframework.data.mongodb.core.aggregation.ComparisonOperators
import org.springframework.data.mongodb.core.query.Criteria

class EqualOperator(
    searchBy: String,
    values: List<Any>
) : SearchOperator(searchBy, values) {

    override fun appendExpression(criteria: Criteria): Criteria = criteria.`in`(values)

    override fun buildExpression(): AggregationExpression =
        ComparisonOperators.Eq.valueOf(searchBy).equalToValue(values[0])
}